+++ frontend/src/domains/user/suggestions/hooks/useSuggestionForm.ts
@@ -1,4 +1,6 @@
+import { postUserFeedback } from '@/apis/userFeedback.api';
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 type SuggestionsFormValues = {
   suggestions: string;
@@ -23,6 +25,7 @@ interface UseSuggestionsFormProps {
 export default function useSuggestionForm({
   initialValues,
 }: UseSuggestionsFormProps) {
+  const navigate = useNavigate();
   const [values, setValues] = useState(initialValues);
 
   const handleSuggestionForm = (input: InputEvent) => {
@@ -35,5 +38,17 @@ export default function useSuggestionForm({
     }
   };
 
-  return { values, handleSuggestionForm };
+  const handleSubmitSuggestions = () => {
+    postUserFeedback({
+      placeId: 1,
+      content: values.suggestions,
+      imageUrl: values.imgSrc,
+      isSecret: values.isSecret,
+      userName: values.userName,
+    });
+    alert('건의사항이 제출되었습니다!');
+    navigate('/');
+  };
+
+  return { values, handleSuggestionForm, handleSubmitSuggestions };
 }
